pyproject.toml
[tool.poetry]
name = "mariner"
version = "0.2.0"
description = "Web application for controlling MSLA 3D Printers based on ChiTu controllers"
license = "MIT"
readme = "README.rst"
authors = [
  "Luiz Ribeiro <luizribeiro@gmail.com>",
]

[tool.poetry.scripts]
mariner = 'mariner.server:main'

[tool.poetry.dependencies]
flask = "2.0.1"
flask-caching = "1.10.1"
Flask-WTF = "0.15.1"
pypng = "0.0.20"
pyre-extensions = "0.0.21"
pyserial = "3.5"
python = "^3.7, !=3.9"
toml = "0.10.2"
typedstruct = "0.1.0"
waitress = "2.0.0"
whitenoise = "5.2.0"

[tool.poetry.dev-dependencies]
black = "20.8b1"
flake8 = "3.9.2"
freezegun = "1.1.0"
green = "3.2.6"
pyexpect = "1.0.21"
pyfakefs = "4.5.0"
pyre-check = "0.9.3"
python-language-server = "0.36.2"
sphinx = "4.0.2"
sphinx-rtd-theme = "0.5.2"

[build-system]
requires = ["poetry>=1.0.10"]
build-backend = "poetry.masonry.api"
